Create corr.R
'directory' is a character vector of length 1 indicating the location of the csv files. 'Threshold' is a numeric vector of length 1 indicating the number of completely observed observations (on all variables) required to compute the correlation between nitrate and sulfate; the default is 0. Returns a numeric vector of correlations. All 332 csv files are found in the specdata folder...