Create cacheSolve.R
The following function returns the inverse of the matrix. It first checks if 
the inverse has already been computed. If so, it gets the result and skips the 
computation. If not, it computes the inverse, sets the value in the cache via
setinverse function. This function assumes that the matrix is always invertible.